ADD: fix setting active bank with domains/https;
